remove pre-5.6 threads leftovers
Time-HiRes has had a min ver of 5.6.0 since release 1.9724 in 2011.
dTHXR and aTHXR are for 5.005 threads compat but they are pointless
since there is usage of "unsafe" dTHX in other places in Time-HiRes
currently. Perl's headers don't define those 2 "R" macros, so they were
defined to >= 5.6.0 threads in ppport.h. By switching to the modern >=5.6.0
macros, ppport.h is only used to allow Time::HiRes to run on old perls, not
the other direction of using ppport.h to allow ancient XS (Time::HiRes in
this case) code to run on newer perls. This patch is split off a p5p patch
that may or may not be accepted to use an empty ppport.h on blead, so this
patch allows HiRes to compile on blead perl with an empty ppport.h.