correct documentation of Time::NVtime to match the implementation
Time::HiRes has always used NV as the return value of the function
pointer stored in Time::NVtime, which is fine when perl is built
with default settings, since NV is then just double.

However, if perl is built with -Duselongdouble, NV becomes long double,
and code that follows the documentation exhibits hard to debug
behaviour, eg. [perl #123879].

Some modules on CPAN that use this interface follow the documentation
and some follow the implementation, changing the implementation would
break modules that follow the implementation and break backward
compatibility, so update the documentation.